+++ app/src/main/AndroidManifest.xml
@@ -4,7 +4,6 @@
 
     <!-- 用户权限 -->
     <uses-permission android:name="android.permission.INTERNET" />
-    <uses-permission android:name="android.permission.MOUNT_UNMOUNT_FILESYSTEMS" />
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" />
@@ -76,12 +75,6 @@
         <!-- 文章详情 -->
         <activity
             android:name=".mvp.ui.activity.NewsDetailsActivity"
-            android:screenOrientation="portrait"
-            android:theme="@style/AppTheme" />
-
-        <!-- 视频详情 -->
-        <activity
-            android:name=".mvp.ui.fragment.VideoDetailsFragment"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme" />
