+++ app/src/main/java/com/stx/xhb/dmgameapp/data/remote/TasksRepositoryProxy.java
@@ -1,6 +1,8 @@
 package com.stx.xhb.dmgameapp.data.remote;
 
 import com.stx.xhb.dmgameapp.config.ApiService;
+import com.stx.xhb.dmgameapp.data.callback.LoadTaskCallback;
+import com.stx.xhb.dmgameapp.http.HttpResultSubscriber;
 import com.stx.xhb.dmgameapp.utils.RequestBodyHelper;
 import com.stx.xhb.dmgameapp.data.TasksDataSource;
 import com.stx.xhb.dmgameapp.data.entity.NewsContent;
@@ -9,12 +11,10 @@ import com.stx.xhb.dmgameapp.http.HttpManager;
 import com.stx.xhb.dmgameapp.http.HttpResult;
 import com.stx.xhb.dmgameapp.http.TransformUtils;
 
-import rx.Subscriber;
 import rx.Subscription;
 
 /**
  * Author: Mr.xiao on 2018/9/5
- *
  * @mail:xhb_199409@163.com
  * @github:https://github.com/xiaohaibin
  * @describe: 数据仓库代理类，用于封装从不同来源处获取数据
@@ -43,42 +43,144 @@ public class TasksRepositoryProxy implements TasksDataSource {
     }
 
     @Override
-    public Subscription getHowNews(int currentPage, Subscriber<? super HttpResult<NewsPageBean>> subscriber) {
+    public Subscription getHowNews(int currentPage, final LoadTaskCallback<NewsPageBean> callback) {
         return HttpManager.getInstance().createService(ApiService.class)
                 .getHotNews(RequestBodyHelper.creatRequestBody(new NewsContent(currentPage)))
                 .compose(TransformUtils.<HttpResult<NewsPageBean>>defaultSchedulers())
-                .subscribe(subscriber);
+                .subscribe(new HttpResultSubscriber<NewsPageBean>() {
+                    @Override
+                    public void onStart() {
+                        callback.onStart();
+                    }
+
+                    @Override
+                    public void onSuccess(NewsPageBean newsPageBean) {
+                        callback.onTaskLoaded(newsPageBean);
+                    }
+
+                    @Override
+                    public void onError(String msg, int code) {
+                        callback.onDataNotAvailable(msg);
+                    }
+
+                    @Override
+                    public void onFinished() {
+                        callback.onCompleted();
+                    }
+                });
     }
 
     @Override
-    public Subscription getNews(int currentPage, Subscriber<? super HttpResult<NewsPageBean>> subscriber) {
+    public Subscription getNews(int currentPage, final LoadTaskCallback<NewsPageBean> callback) {
         return HttpManager.getInstance().createService(ApiService.class)
                 .getNews(RequestBodyHelper.creatRequestBody(new NewsContent(currentPage)))
                 .compose(TransformUtils.<HttpResult<NewsPageBean>>defaultSchedulers())
-                .subscribe(subscriber);
+                .subscribe(new HttpResultSubscriber<NewsPageBean>() {
+                    @Override
+                    public void onStart() {
+                        callback.onStart();
+                    }
+
+                    @Override
+                    public void onSuccess(NewsPageBean newsPageBean) {
+                        callback.onTaskLoaded(newsPageBean);
+                    }
+
+                    @Override
+                    public void onError(String msg, int code) {
+                        callback.onDataNotAvailable(msg);
+                    }
+
+                    @Override
+                    public void onFinished() {
+                        callback.onCompleted();
+                    }
+                });
     }
 
     @Override
-    public Subscription getOriginalPage(int currentPage, Subscriber<? super HttpResult<NewsPageBean>> subscriber) {
+    public Subscription getOriginalPage(int currentPage, final LoadTaskCallback<NewsPageBean> callback) {
         return HttpManager.getInstance().createService(ApiService.class)
                 .getOriginalPage(RequestBodyHelper.creatRequestBody(new NewsContent(currentPage)))
                 .compose(TransformUtils.<HttpResult<NewsPageBean>>defaultSchedulers())
-                .subscribe(subscriber);
+                .subscribe(new HttpResultSubscriber<NewsPageBean>() {
+
+                    @Override
+                    public void onStart() {
+                        callback.onStart();
+                    }
+
+                    @Override
+                    public void onSuccess(NewsPageBean newsPageBean) {
+                        callback.onTaskLoaded(newsPageBean);
+                    }
+
+                    @Override
+                    public void onError(String msg, int code) {
+                        callback.onDataNotAvailable(msg);
+                    }
+
+                    @Override
+                    public void onFinished() {
+                        callback.onCompleted();
+                    }
+                });
     }
 
     @Override
-    public Subscription getVideoPage(int currentPage, Subscriber<? super HttpResult<NewsPageBean>> subscriber) {
+    public Subscription getVideoPage(int currentPage, final LoadTaskCallback<NewsPageBean> callback) {
         return HttpManager.getInstance().createService(ApiService.class)
                 .getVideoPage(RequestBodyHelper.creatRequestBody(new NewsContent(currentPage)))
                 .compose(TransformUtils.<HttpResult<NewsPageBean>>defaultSchedulers())
-                .subscribe(subscriber);
+                .subscribe(new HttpResultSubscriber<NewsPageBean>() {
+
+                    @Override
+                    public void onStart() {
+                        callback.onStart();
+                    }
+
+                    @Override
+                    public void onSuccess(NewsPageBean newsPageBean) {
+                        callback.onTaskLoaded(newsPageBean);
+                    }
+
+                    @Override
+                    public void onError(String msg, int code) {
+                        callback.onDataNotAvailable(msg);
+                    }
+
+                    @Override
+                    public void onFinished() {
+                        callback.onCompleted();
+                    }
+                });
     }
 
     @Override
-    public Subscription getAmusePage(int currentPage, Subscriber<? super HttpResult<NewsPageBean>> subscriber) {
+    public Subscription getAmusePage(int currentPage, final LoadTaskCallback<NewsPageBean> callback) {
         return HttpManager.getInstance().createService(ApiService.class)
                 .getAmusePage(RequestBodyHelper.creatRequestBody(new NewsContent(currentPage)))
                 .compose(TransformUtils.<HttpResult<NewsPageBean>>defaultSchedulers())
-                .subscribe(subscriber);
+                .subscribe(new HttpResultSubscriber<NewsPageBean>() {
+                    @Override
+                    public void onStart() {
+                        callback.onStart();
+                    }
+
+                    @Override
+                    public void onSuccess(NewsPageBean newsPageBean) {
+                        callback.onTaskLoaded(newsPageBean);
+                    }
+
+                    @Override
+                    public void onError(String msg, int code) {
+                        callback.onDataNotAvailable(msg);
+                    }
+
+                    @Override
+                    public void onFinished() {
+                        callback.onCompleted();
+                    }
+                });
     }
 }
