+++ app/src/main/java/com/stx/xhb/dmgameapp/http/BaseResponse.java
@@ -8,7 +8,7 @@ import java.io.Serializable;
  * @author : Mr.xiao on 2017/3/15
  * @mail:xhb_199409@163.com
  * @github:https://github.com/xiaohaibin
- * @describe:
+ * @describe: BaseResponse
  */
 public class BaseResponse implements Serializable{
