+++ app/src/main/java/com/stx/xhb/dmgameapp/mvp/ui/adapter/ForumDetailsListAdapter.java
@@ -1,4 +1,4 @@
-package com.stx.xhb.dmgameapp.adapter;
+package com.stx.xhb.dmgameapp.mvp.ui.adapter;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -7,8 +7,8 @@ import android.view.ViewGroup;
 import com.jude.easyrecyclerview.adapter.BaseViewHolder;
 import com.jude.easyrecyclerview.adapter.RecyclerArrayAdapter;
 import com.stx.xhb.dmgameapp.R;
-import com.stx.xhb.dmgameapp.adapter.viewholder.ForumDetailsListViewHodler;
 import com.stx.xhb.dmgameapp.data.entity.ForumListBean;
+import com.stx.xhb.dmgameapp.mvp.ui.adapter.viewholder.ForumDetailsListViewHodler;
 
 
 /**
