+++ app/src/main/java/com/stx/xhb/dmgameapp/mvp/ui/adapter/ForumListAdapter.java
@@ -1,4 +1,4 @@
-package com.stx.xhb.dmgameapp.adapter;
+package com.stx.xhb.dmgameapp.mvp.ui.adapter;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -7,28 +7,28 @@ import android.view.ViewGroup;
 import com.jude.easyrecyclerview.adapter.BaseViewHolder;
 import com.jude.easyrecyclerview.adapter.RecyclerArrayAdapter;
 import com.stx.xhb.dmgameapp.R;
-import com.stx.xhb.dmgameapp.adapter.viewholder.VideoDetailsItemViewHolder;
-import com.stx.xhb.dmgameapp.data.entity.VideoListBean;
+import com.stx.xhb.dmgameapp.data.entity.ForumBean;
+import com.stx.xhb.dmgameapp.mvp.ui.adapter.viewholder.ForumListViewHodler;
 
 /**
  * Author：xiaohaibin
- * Time：2017/9/19
+ * Time：2017/9/21
  * Emil：xhb_199409@163.com
  * Github：https://github.com/xiaohaibin/
- * Describe：视频详情列表适配器
+ * Describe：
  */
 
-public class VideoDetailsListAdapter extends RecyclerArrayAdapter<VideoListBean.VideoBean> {
+public class ForumListAdapter extends RecyclerArrayAdapter<ForumBean> {
 
-    private LayoutInflater mLayoutInflater;
+    private final LayoutInflater mLayoutInflater;
 
-    public VideoDetailsListAdapter(Context context) {
+    public ForumListAdapter(Context context) {
         super(context);
         mLayoutInflater = LayoutInflater.from(context);
     }
 
     @Override
     public BaseViewHolder OnCreateViewHolder(ViewGroup parent, int viewType) {
-        return new VideoDetailsItemViewHolder(mLayoutInflater.inflate(R.layout.list_item_video_details, parent, false));
+        return new ForumListViewHodler(mLayoutInflater.inflate(R.layout.list_item_forum, parent, false));
     }
 }
