+++ app/src/main/java/com/stx/xhb/dmgameapp/mvp/ui/adapter/VideoDetailsListAdapter.java
@@ -1,4 +1,4 @@
-package com.stx.xhb.dmgameapp.adapter;
+package com.stx.xhb.dmgameapp.mvp.ui.adapter;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -7,29 +7,28 @@ import android.view.ViewGroup;
 import com.jude.easyrecyclerview.adapter.BaseViewHolder;
 import com.jude.easyrecyclerview.adapter.RecyclerArrayAdapter;
 import com.stx.xhb.dmgameapp.R;
-import com.stx.xhb.dmgameapp.adapter.viewholder.VideoListViewHolder;
-import com.stx.xhb.dmgameapp.data.entity.NewsPageBean;
 import com.stx.xhb.dmgameapp.data.entity.VideoListBean;
+import com.stx.xhb.dmgameapp.mvp.ui.adapter.viewholder.VideoDetailsItemViewHolder;
 
 /**
  * Author：xiaohaibin
  * Time：2017/9/19
  * Emil：xhb_199409@163.com
  * Github：https://github.com/xiaohaibin/
- * Describe：视频列表适配器
+ * Describe：视频详情列表适配器
  */
 
-public class VideoListAdapter extends RecyclerArrayAdapter<NewsPageBean.ListBean> {
+public class VideoDetailsListAdapter extends RecyclerArrayAdapter<VideoListBean.VideoBean> {
 
     private LayoutInflater mLayoutInflater;
 
-    public VideoListAdapter(Context context) {
+    public VideoDetailsListAdapter(Context context) {
         super(context);
         mLayoutInflater = LayoutInflater.from(context);
     }
 
     @Override
     public BaseViewHolder OnCreateViewHolder(ViewGroup parent, int viewType) {
-        return new VideoListViewHolder(mLayoutInflater.inflate(R.layout.list_item_video, parent, false));
+        return new VideoDetailsItemViewHolder(mLayoutInflater.inflate(R.layout.list_item_video_details, parent, false));
     }
 }
