+++ app/src/main/java/com/stx/xhb/dmgameapp/mvp/ui/adapter/viewholder/NewsCommonViewHolder.java
@@ -1,4 +1,4 @@
-package com.stx.xhb.dmgameapp.adapter.viewholder;
+package com.stx.xhb.dmgameapp.mvp.ui.adapter.viewholder;
 
 import android.view.View;
 import android.widget.ImageView;
@@ -9,7 +9,6 @@ import com.jude.easyrecyclerview.adapter.BaseViewHolder;
 import com.stx.core.utils.DateUtils;
 import com.stx.xhb.dmgameapp.R;
 import com.stx.xhb.dmgameapp.data.entity.NewsPageBean;
-import com.stx.xhb.dmgameapp.data.entity.VideoListBean;
 import com.stx.xhb.dmgameapp.mvp.ui.activity.NewsDetailsActivity;
 import com.stx.xhb.dmgameapp.mvp.ui.activity.WebDetailsActivity;
 
@@ -17,35 +16,35 @@ import butterknife.Bind;
 import butterknife.ButterKnife;
 
 /**
- * @Author：xiaohaibin
- * @Time：2017/9/19
- * @Emil：xhb_199409@163.com
+ * @author：xiaohaibin
+ * @time：2017/9/18
+ * @emil：xhb_199409@163.com
  * @Github：https://github.com/xiaohaibin/
  * @Describe：
  */
-public class VideoListViewHolder extends BaseViewHolder<NewsPageBean.ListBean> {
 
-    @Bind(R.id.iv_video_img)
-    ImageView mIvVideoImg;
-    @Bind(R.id.tv_video_title)
-    TextView mTvVideoTitle;
-    @Bind(R.id.tv_video_time)
-    TextView mTvVideoTime;
-    @Bind(R.id.tv_play_count)
-    TextView mTvPlayCount;
+public class NewsCommonViewHolder extends BaseViewHolder<NewsPageBean.ListBean> {
+    @Bind(R.id.title)
+    TextView mTitle;
+    @Bind(R.id.date)
+    TextView mDate;
+    @Bind(R.id.iv)
+    ImageView mIv;
+    @Bind(R.id.author)
+    TextView tvAuthor;
+    private String imgUrl = "";
 
-    public VideoListViewHolder(View itemView) {
+    public NewsCommonViewHolder(View itemView) {
         super(itemView);
         ButterKnife.bind(this, itemView);
     }
 
     @Override
     public void setData(final NewsPageBean.ListBean data) {
-        super.setData(data);
-        mTvVideoTitle.setText(data.getTitle());
-        mTvVideoTime.setText(DateUtils.getFriendlyTime(data.getPubdate_at()+"000"));
-        mTvPlayCount.setText(String.valueOf(data.getClick() + "次播放"));
-        Glide.with(getContext()).load(data.getLitpic()).into(mIvVideoImg);
+        mTitle.setText(data.getTitle());
+        tvAuthor.setText(data.getUser().getNickname());
+        mDate.setText(DateUtils.getFriendlyTime(data.getPubdate_at()+"000"));
+        Glide.with(getContext()).load(data.getLitpic()).into(mIv);
         itemView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
