+++ app/src/main/java/com/stx/xhb/dmgameapp/mvp/ui/adapter/viewholder/VideoDetailsItemViewHolder.java
@@ -1,4 +1,4 @@
-package com.stx.xhb.dmgameapp.adapter.viewholder;
+package com.stx.xhb.dmgameapp.mvp.ui.adapter.viewholder;
 
 import android.view.View;
 import android.widget.ImageView;
@@ -6,11 +6,9 @@ import android.widget.TextView;
 
 import com.bumptech.glide.Glide;
 import com.jude.easyrecyclerview.adapter.BaseViewHolder;
-import com.stx.core.utils.DateUtils;
 import com.stx.xhb.dmgameapp.R;
-import com.stx.xhb.dmgameapp.data.entity.NewsPageBean;
+import com.stx.xhb.dmgameapp.data.entity.VideoListBean;
 import com.stx.xhb.dmgameapp.mvp.ui.activity.NewsDetailsActivity;
-import com.stx.xhb.dmgameapp.mvp.ui.activity.WebDetailsActivity;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
@@ -23,32 +21,32 @@ import butterknife.ButterKnife;
  * @Describe：
  */
 
-public class NewsCommonViewHolder extends BaseViewHolder<NewsPageBean.ListBean> {
+public class VideoDetailsItemViewHolder extends BaseViewHolder<VideoListBean.VideoBean> {
+
     @Bind(R.id.title)
     TextView mTitle;
     @Bind(R.id.date)
     TextView mDate;
+    @Bind(R.id.tv_play_count)
+    TextView mTvPlayCount;
     @Bind(R.id.iv)
     ImageView mIv;
-    @Bind(R.id.author)
-    TextView tvAuthor;
-    private String imgUrl = "";
 
-    public NewsCommonViewHolder(View itemView) {
+    public VideoDetailsItemViewHolder(View itemView) {
         super(itemView);
         ButterKnife.bind(this, itemView);
     }
 
     @Override
-    public void setData(final NewsPageBean.ListBean data) {
+    public void setData(final VideoListBean.VideoBean data) {
         mTitle.setText(data.getTitle());
-        tvAuthor.setText(data.getUser().getNickname());
-        mDate.setText(DateUtils.getFriendlyTime(data.getPubdate_at()+"000"));
-        Glide.with(getContext()).load(data.getLitpic()).into(mIv);
+        mDate.setText(data.getSenddate());
+        mTvPlayCount.setText(data.getClick() + "次播放");
+        Glide.with(getContext()).load(data.getVideopic()).into(mIv);
         itemView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                WebDetailsActivity.start(getContext(),data.getWebviewurl(),data.getTitle(),data.getLitpic());
+                NewsDetailsActivity.start(getContext(),data.getVideourl(),data.getId(),data.getTitle(),data.getVideopic(),true);
             }
         });
     }
