+++ app/src/main/java/com/stx/xhb/dmgameapp/mvp/ui/fragment/CommentListFragment.java
@@ -13,10 +13,10 @@ import com.jude.easyrecyclerview.decoration.DividerDecoration;
 import com.stx.core.base.BaseMvpFragment;
 import com.stx.core.utils.ScreenUtil;
 import com.stx.xhb.dmgameapp.R;
-import com.stx.xhb.dmgameapp.adapter.CommentListAdapter;
 import com.stx.xhb.dmgameapp.data.entity.CommentListBean;
 import com.stx.xhb.dmgameapp.mvp.contract.GetCommentListContract;
 import com.stx.xhb.dmgameapp.mvp.presenter.GetCommentListPresenter;
+import com.stx.xhb.dmgameapp.mvp.ui.adapter.CommentListAdapter;
 import com.stx.xhb.dmgameapp.utils.ToastUtil;
 import com.stx.xhb.dmgameapp.widget.widget.CustomTitlebar;
