+++ app/src/main/java/com/stx/xhb/dmgameapp/mvp/ui/main/GameFragment.java
@@ -9,7 +9,7 @@ import android.widget.TextView;
 import com.stx.core.base.BaseFragment;
 import com.stx.core.base.BaseMvpFragment;
 import com.stx.xhb.dmgameapp.R;
-import com.stx.xhb.dmgameapp.adapter.GameViewPagerFragmentAdapter;
+import com.stx.xhb.dmgameapp.mvp.ui.adapter.GameViewPagerFragmentAdapter;
 import com.stx.xhb.dmgameapp.mvp.ui.news.HotNewsFragment;
 
 import java.util.ArrayList;
